Updated the Navigation Drawers to the better working type and deleted old references
