+++ src/js/NavigationDrawers/NavigationDrawer.js
@@ -1,49 +1,32 @@
 import React, { Component, PropTypes } from 'react';
 import PureRenderMixin from 'react-addons-pure-render-mixin';
+import CSSTransitionGroup from 'react-addons-css-transition-group';
 import classnames from 'classnames';
 
-import { setOverflow, isTouchDevice } from '../utils';
-import Divider from '../Dividers';
-import { List, ListItem } from '../Lists';
-import Subheader from '../Subheaders';
+import Drawer from './Drawer';
+import DrawerToolbar from './DrawerToolbar';
 import Overlay from '../Transitions/Overlay';
 
-import NavigationDrawerHeader from './NavigationDrawerHeader';
-import NavigationDrawerToolbar from './NavigationDrawerToolbar';
-
-const deprecate = (props, propName, component, to) => {
-  if(typeof props[propName] !== 'undefined') {
-    console.error(
-      `The prop '${propName}' has been renamed to '${to}' ` +
-      'for the next release. Check the upgrade docs at ' +
-      'https://github.com/mlaursen/react-md/tree/master/docs/UpgradeGuide.md ' +
-      'for more information.'
-    );
-  }
-};
-
-const removed = (props, propName) => {
-  if(typeof props[propName] !== 'undefined') {
-    console.error(
-      `The prop '${propName}' has been removed for the next release. Check ` +
-      'the upgrade docs at ' +
-      'https://github.com/mlaursen/react-md/tree/master/docs/UpgradeGuide.md ' +
-      'for more information.'
-    );
-  }
-};
-
 /**
- * Navigation drawers are an excellent component to use to set up the initial
- * layout of your application. This component combines a Navigation drawer
+ * Navigation Drawers are an excellent component to use to set up the initial
+ * layout of your application. This component combines a Drawer
  * (a sidebar of nav items), an app bar, and displays any additional content.
+ *
+ * The `NavigationDrawer` component is customizable to have different
+ * display states for mobile, tablet, and desktop displays.
  */
 export default class NavigationDrawer extends Component {
   constructor(props) {
     super(props);
 
     this.shouldComponentUpdate = PureRenderMixin.shouldComponentUpdate.bind(this);
-    this.state = { touch: false };
+    this.state = {
+      isOpen: typeof props.initiallyOpen !== 'undefined' ? props.initiallyOpen : true,
+      mobile: false,
+      tablet: false,
+      desktop: true,
+      drawerType: props.desktopDrawerType,
+    };
   }
 
   static DrawerType = {
@@ -59,36 +42,24 @@ export default class NavigationDrawer extends Component {
 
   static propTypes = {
     /**
-     * An optional style to apply to the navigation drawer itself.
+     * An optional style to apply to the entire container.
      */
-    style: (props, propName, component) => {
-      deprecate(props, propName, component, 'drawerStyle');
-      return PropTypes.object(props, propName, component);
-    },
+    style: PropTypes.object,
 
     /**
-     * An optional className to apply to the navigation drawer iteself.
+     * An optional className to apply to the entire container.
      */
-    className: (props, propName, component) => {
-      deprecate(props, propName, component, 'drawerClassName');
-      return PropTypes.string(props, propName, component);
-    },
+    className: PropTypes.string,
 
     /**
-     * An optional style to apply to the navigation drawer container.
+     * An optional style to apply to the drawer.
      */
-    containerStyle: (props, propName, component) => {
-      deprecate(props, propName, component, 'style');
-      return PropTypes.object(props, propName, component);
-    },
+    drawerStyle: PropTypes.object,
 
     /**
-     * An optional className to apply to the navigation drawer container.
+     * An optional className to apply to the drawer.
      */
-    containerClassName: (props, propName, component) => {
-      deprecate(props, propName, component, 'className');
-      return PropTypes.string(props, propName, component);
-    },
+    drawerClassName: PropTypes.string,
 
     /**
      * An optional style to apply to the main content.
@@ -111,25 +82,91 @@ export default class NavigationDrawer extends Component {
     toolbarClassName: PropTypes.string,
 
     /**
-     * Boolean if the navigation drawer is currently open.
-     */
-    isOpen: (props, propName, component) => {
-      removed(props, propName);
-      return PropTypes.bool.isRequired(props, propName, component);
-    },
-
-    /**
      * An optional title to display in the navigation drawer header.
      */
-    title: (props, propName, component) => {
-      deprecate(props, propName, component, 'drawerTitle');
-      return PropTypes.string(props, propName, component);
-    },
+    drawerTitle: PropTypes.string,
+
+    /**
+     * Any additional children you want to display in the navigation drawer header after
+     * the optional title.
+     */
+    drawerChildren: PropTypes.node,
+
+    /**
+     * An optional title to display in the toolbar.
+     */
+    toolbarTitle: PropTypes.string,
+
+    /**
+     * Any additional children you want to display in the main toolbar after the menu
+     * button (for persistent drawers) and the optional title.
+     */
+    toolbarChildren: PropTypes.node,
 
     /**
      * The main content to display.
      */
     children: PropTypes.node,
+
+    /**
+     * An optional boolean if the drawer is open for persistent or temporary
+     * drawers.
+     */
+    initiallyOpen: PropTypes.bool,
+
+    /**
+     * The `DrawerType` to use for mobile devices. If the `mobileMinWidth` prop
+     * matches or the `tabletMinWidth` prop matches and the orientation is portrait,
+     * this drawer type will be used.
+     */
+    mobileDrawerType: PropTypes.oneOf([
+      NavigationDrawer.DrawerType.TEMPORARY,
+      NavigationDrawer.DrawerType.TEMPORARY_MINI,
+    ]).isRequired,
+
+    /**
+     * The `DrawerType` to use for landscape oriented tablets. If the
+     * `tabletMinWidth` prop matches and the orientation is landscape, this drawer
+     * type will be used.
+     */
+    tabletDrawerType: PropTypes.oneOf([
+      NavigationDrawer.DrawerType.FULL_HEIGHT,
+      NavigationDrawer.DrawerType.CLIPPED,
+      NavigationDrawer.DrawerType.FLOATING,
+      NavigationDrawer.DrawerType.PERSISTENT,
+      NavigationDrawer.DrawerType.PERSISTENT_MINI,
+      NavigationDrawer.DrawerType.TEMPORARY,
+      NavigationDrawer.DrawerType.TEMPORARY_MINI,
+    ]).isRequired,
+
+    /**
+     * The `DrawerType` to use for desktop displays. If the `desktopMinWidth`
+     * prop matches, this drawer type will be used.
+     */
+    desktopDrawerType: PropTypes.oneOf([
+      NavigationDrawer.DrawerType.FULL_HEIGHT,
+      NavigationDrawer.DrawerType.CLIPPED,
+      NavigationDrawer.DrawerType.FLOATING,
+      NavigationDrawer.DrawerType.PERSISTENT,
+      NavigationDrawer.DrawerType.PERSISTENT_MINI,
+      NavigationDrawer.DrawerType.TEMPORARY,
+      NavigationDrawer.DrawerType.TEMPORARY_MINI,
+    ]).isRequired,
+
+    /**
+     * The min width to use for a mobile media query.
+     */
+    mobileMinWidth: PropTypes.number.isRequired,
+
+    /**
+     * The min width to use for a tablet media query.
+     */
+    tabletMinWidth: PropTypes.number.isRequired,
+
+    /**
+     * The min width to use for a desktop media query.
+     */
+    desktopMinWidth: PropTypes.number.isRequired,
 
     /**
      * The icon className to use for the main menu icon.
@@ -152,31 +189,20 @@ export default class NavigationDrawer extends Component {
     closeIconChildren: PropTypes.node,
 
     /**
-     * A function to call that will open the navigation drawer. This will
-     * be applied to the main menu button.
+     * Boolean if the sliding drawer should automatically close when a nav item
+     * is clicked for a temporary or temporary mini drawer.
      */
-    openDrawer: (props, propName, component) => {
-      removed(props, propName);
-      return PropTypes.func(props, propName, component);
-    },
+    autoclose: PropTypes.bool.isRequired,
 
     /**
-     * A function to call that will close the navigation drawer. If the drawer
-     * is persistent or persistent mini, this function is required.
+     * An optional function to call when the sliding drawer's open state changes.
+     * This can happen from rotating a tablet, resizing the browser window,
+     * or toggling the drawer for persistent and temporary versions. The
+     * `onDrawerChange` prop will be given the new open state.
+     *
+     * `onDrawerChange(isOpen)`
      */
-    closeDrawer: (props, propName, component) => {
-      removed(props, propName);
-
-      const { PERSISTENT, PERSISTENT_MINI } = NavigationDrawer.DrawerType;
-      const { drawerType } = props;
-      if(drawerType !== PERSISTENT && drawerType !== PERSISTENT_MINI) {
-        return;
-      }
-
-      return PropTypes.func.isRequired(props, propName, component)
-        || PropTypes.string(props, 'closeIconClassName', component)
-        || PropTypes.node(props, 'closeIconChildren', component);
-    },
+    onDrawerChange: PropTypes.func,
 
     /**
      * A list of items to render in the navigation drawer. If an item
@@ -186,7 +212,17 @@ export default class NavigationDrawer extends Component {
      * ##### Additional Info
      */
     navItems: PropTypes.arrayOf(PropTypes.oneOfType([
-      PropTypes.node,
+      /**
+       * Any react component you want to render instead of using props
+       * to generate a component.
+       */
+      PropTypes.element,
+
+      /**
+       * An object of props to use that will generate either a `ListItem`, `Divider`,
+       * or a `Subheader` component. Any props not listed will be passed
+       * to the component.
+       */
       PropTypes.shape({
         /**
          * Boolean if this item should be rendered as a divider.
@@ -214,183 +250,267 @@ export default class NavigationDrawer extends Component {
     ])).isRequired,
 
     /**
-     * The drawer type to use.
+     * The transition name to use for the drawer sliding in and out of view.
      */
-    drawerType: (props, propName, component) => {
-      removed(props, propName);
-      return PropTypes.oneOf(Object.keys(NavigationDrawer.DrawerType).map(k => NavigationDrawer.DrawerType[k]))(props, propName, component);
-    },
+    transitionName: PropTypes.string.isRequired,
 
     /**
-     * Any additional children to display in the main toolbar.
+     * The transition enter duration for the drawer sliding in to view.
      */
-    toolbarChildren: PropTypes.node,
+    transitionEnterTimeout: PropTypes.number.isRequired,
 
     /**
-     * An optional title to display in the main toolbar.
+     * The transition leave duration for the drawer sliding out of view.
      */
-    toolbarTitle: PropTypes.string,
+    transitionLeaveTimeout: PropTypes.number.isRequired,
 
     /**
-     * An optional header to display in the navigation drawer. This will be rendered
-     * before the `navItems`. This will replace the generated header if you gave the
-     * `title` prop.
+     * An optional transition to use when the content changes. To disable this transition,
+     * set the enter and leave timeouts to `0`, `null`, or `undefined`. Make sure your
+     * child has a new key as well.
      */
-    navHeader: (props, propName, component) => {
-      removed(props, propName);
-      return PropTypes.node(props, propName, component);
-    },
+    contentTransitionName: PropTypes.string.isRequired,
 
     /**
-     * Any additional children to display in the navigation drawer's header after the
-     * `title`.
+     * The transition enter timeout for when the main content changes. Settings this value
+     * to `0`, `null`, or `undefined` will disable the enter transition.
      */
-    navHeaderChildren: (props, propName, component) => {
-      deprecate(props, propName, component, 'drawerChildren');
-      return PropTypes.node(props, propName, component);
-    },
+    contentTransitionEnterTimeout: PropTypes.number,
+
+    /**
+     * The transition leave timeout for when the main content changes. Settings this value
+     * to `0`, `null`, or `undefined` will disable the leave transition.
+     */
+    contentTransitionLeaveTimeout: PropTypes.number,
   };
 
   static defaultProps = {
-    drawerType: NavigationDrawer.DrawerType.FULL_HEIGHT,
+    mobileDrawerType: NavigationDrawer.DrawerType.TEMPORARY,
+    tabletDrawerType: NavigationDrawer.DrawerType.PERSISTENT,
+    desktopDrawerType: NavigationDrawer.DrawerType.FULL_HEIGHT,
+    mobileMinWidth: 320,
+    tabletMinWidth: 768,
+    desktopMinWidth: 1025,
     menuIconChildren: 'menu',
+    autoclose: true,
     closeIconChildren: 'keyboard_arrow_left',
+    transitionName: 'drawer',
+    transitionEnterTimeout: 300,
+    transitionLeaveTimeout: 300,
+    contentTransitionName: 'cross-fade',
+    contentTransitionEnterTimeout: 300,
   };
 
   componentDidMount() {
-    this.setState({ touch: isTouchDevice() }); // eslint-disable-line
+    window.addEventListener('resize', this._updateMedia);
+    this._updateMedia();
   }
 
-  componentWillUpdate(nextProps) {
-    if(nextProps.isOpen === this.props.isOpen && nextProps.drawerType === this.props.drawerType) {
-      return;
-    }
+  componentWillReceiveProps(nextProps) {
+    const {
+      mobileMinWidth,
+      mobileDrawerType,
+      tabletMinWidth,
+      tabletDrawerType,
+      desktopMinWidth,
+      desktopDrawerType,
+    } = this.props;
 
-    const temps = [
-      NavigationDrawer.DrawerType.TEMPORARY,
-      NavigationDrawer.DrawerType.TEMPORARY_MINI,
-    ];
-
-    const isNextTemp = temps.indexOf(nextProps.drawerType) !== -1;
-    const isCurrTemp = temps.indexOf(this.props.drawerType) !== -1;
-    if(!isNextTemp && !isCurrTemp) { return; }
-    if((isCurrTemp && !isNextTemp) || (!isCurrTemp && !isNextTemp)) {
-      setOverflow(false);
-    } else {
-      setOverflow(nextProps.isOpen);
+    if(nextProps.mobileMinWidth !== mobileMinWidth || nextProps.mobileDrawerType !== mobileDrawerType
+      || nextProps.tabletMinWidth !== tabletMinWidth || nextProps.tabletDrawerType !== tabletDrawerType
+      || nextProps.desktopMinWidth !== desktopMinWidth || nextProps.desktopDrawerType !== desktopDrawerType) {
+      this._updateDrawerType(nextProps);
     }
   }
 
-  mapItemsToComponents = (mini, active, item, key) => {
-    if(React.isValidElement(item)) {
-      return item;
+  componentWillUpdate(nextProps, nextState) {
+    if(this.state.isOpen !== nextState.isOpen && nextProps.onDrawerChange) {
+      nextProps.onDrawerChange(nextState.isOpen);
+    }
+  }
+
+  componentWillUnmount() {
+    window.removeEventListener('resize', this._updateMedia);
+  }
+
+  _matches = (min, max, orientation) => {
+    let media = 'only screen';
+    if(orientation) {
+      media += ` and (orientation: ${orientation})`;
     }
 
-    const { divider, subheader, primaryText, nestedItems, ...itemProps } = item;
-    let component;
-    if(divider) {
-      component = Divider;
-    } else if(subheader) {
-      component = Subheader;
+    media += ` and (min-width: ${min}px)`;
+
+    if(max) {
+      media += ` and (max-width: ${max}px)`;
+    }
+
+    return window.matchMedia(media).matches;
+  };
+
+  _updateDrawerType = (props) => {
+    const {
+      mobileMinWidth,
+      mobileDrawerType,
+      tabletMinWidth,
+      tabletDrawerType,
+      desktopMinWidth,
+      desktopDrawerType,
+    } = props;
+
+    const nextState = {
+      mobile: this._matches(mobileMinWidth, desktopMinWidth - 1),
+      tablet: this._matches(tabletMinWidth, desktopMinWidth - 1, 'landscape'),
+      desktop: this._matches(desktopMinWidth),
+    };
+
+    if(nextState.tablet) {
+      nextState.drawerType = tabletDrawerType;
+    } else if(nextState.mobile) {
+      nextState.drawerType = mobileDrawerType;
     } else {
-      component = ListItem;
+      nextState.drawerType = desktopDrawerType;
     }
 
-    const props = Object.assign({}, itemProps, {
-      key: item.key || key,
-      nestedItems: nestedItems && nestedItems.map(this.mapItemsToComponents.bind(this, mini, active)),
-    });
-    if(!mini || (mini && active)) {
-      props.primaryText = primaryText;
+    if(this._isFullHeight(nextState.drawerType) !== this._isFullHeight(this.state.drawerType)) {
+      nextState.isOpen = this._isFullHeight(nextState.drawerType);
     }
 
-    return React.createElement(component, props);
+    this.setState(nextState);
+  };
+
+  _updateMedia = () => {
+    this._updateDrawerType(this.props);
+  };
+
+  openDrawer = () => {
+    this.setState({ isOpen: true });
+  };
+
+  closeDrawer = () => {
+    this.setState({ isOpen: false });
+  };
+
+  _isFullHeight = (drawerType) => {
+    const { FULL_HEIGHT, CLIPPED, FLOATING } = NavigationDrawer.DrawerType;
+    return [FULL_HEIGHT, CLIPPED, FLOATING].indexOf(drawerType) !== -1;
+  };
+
+  _isPersistent = (drawerType) => {
+    const { PERSISTENT, PERSISTENT_MINI } = NavigationDrawer.DrawerType;
+    return [PERSISTENT, PERSISTENT_MINI].indexOf(drawerType) !== -1;
+  };
+
+  _isTemporary = (drawerType) => {
+    const { TEMPORARY, TEMPORARY_MINI } = NavigationDrawer.DrawerType;
+    return [TEMPORARY, TEMPORARY_MINI].indexOf(drawerType) !== -1;
+  };
+
+  _isMini = (drawerType) => {
+    const { PERSISTENT_MINI, TEMPORARY_MINI } = NavigationDrawer.DrawerType;
+    return [PERSISTENT_MINI, TEMPORARY_MINI].indexOf(drawerType) !== -1;
   };
 
   render() {
     const {
-      isOpen,
-      title,
-      toolbarTitle,
       style,
       className,
-      containerStyle,
-      containerClassName,
+      drawerStyle,
+      drawerClassName,
       contentStyle,
       contentClassName,
       toolbarStyle,
       toolbarClassName,
-      menuIconClassName,
+      drawerTitle,
+      drawerChildren,
       menuIconChildren,
-      children,
-      openDrawer,
-      closeDrawer,
-      closeIconClassName,
+      menuIconClassName,
       closeIconChildren,
-      navItems,
-      drawerType,
+      closeIconClassName,
+      toolbarTitle,
       toolbarChildren,
-      navHeader,
-      navHeaderChildren,
+      autoclose,
+      navItems,
+      children,
+      transitionName,
+      transitionEnterTimeout,
+      transitionLeaveTimeout,
+      contentTransitionName,
+      contentTransitionEnterTimeout,
+      contentTransitionLeaveTimeout,
     } = this.props;
 
-    const { touch } = this.state;
-    const { PERSISTENT, PERSISTENT_MINI, TEMPORARY, TEMPORARY_MINI } = NavigationDrawer.DrawerType;
+    const { isOpen, drawerType } = this.state;
+    const persistent = this._isPersistent(drawerType);
+    const temporary = this._isTemporary(drawerType);
+    const mini = this._isMini(drawerType);
 
-    const mini = drawerType === PERSISTENT_MINI || drawerType === TEMPORARY_MINI;
-    const persistent = drawerType === PERSISTENT_MINI || drawerType === PERSISTENT;
-    const temporary = drawerType === TEMPORARY || drawerType === TEMPORARY_MINI;
-    const active = isOpen || (!temporary && (!persistent && !touch));
-
-    let nav;
-    if(active || mini) {
-      const navigationItems = navItems.map(this.mapItemsToComponents.bind(this, mini, active));
-
-      nav = <List>{navigationItems}</List>;
-    }
-
-    let header;
-    if(navHeader) {
-      header = React.cloneElement(navHeader, { persistent });
-    } else if((active && mini) || !mini) {
-      header = (
-        <NavigationDrawerHeader
-          title={title}
-          closeDrawer={closeDrawer}
+    let drawer, overlay;
+    if(isOpen || mini) {
+      drawer = (
+        <Drawer
+          key="drawer"
+          style={drawerStyle}
+          className={drawerClassName}
+          autoclose={autoclose}
+          title={drawerTitle}
+          closeDrawer={this.closeDrawer}
           closeIconChildren={closeIconChildren}
           closeIconClassName={closeIconClassName}
+          temporary={temporary}
           persistent={persistent}
-          children={navHeaderChildren}
+          navItems={navItems}
+          drawerType={drawerType}
+          children={drawerChildren}
+          mini={mini}
+          isOpen={isOpen}
         />
       );
     }
 
-    const conditionalClassNames = { active };
+    if(temporary) {
+      overlay = <Overlay isOpen={isOpen} onClick={this.closeDrawer} />;
+    }
 
     return (
-      <div className={classnames('md-navigation-drawer-container', containerClassName, conditionalClassNames)} style={containerStyle}>
-        <nav className={classnames('md-navigation-drawer', className, drawerType, conditionalClassNames)} style={style}>
-          {header}
-          {nav}
-        </nav>
-        <div className={classnames('md-navigation-drawer-content', contentClassName, drawerType, conditionalClassNames)} style={contentStyle}>
-          <NavigationDrawerToolbar
-            className={classnames(toolbarClassName, drawerType, conditionalClassNames)}
-            temporary={temporary}
+      <CSSTransitionGroup
+        style={style}
+        className={classnames('md-navigation-drawer-container', className)}
+        component="div"
+        transitionName={transitionName}
+        transitionEnterTimeout={transitionEnterTimeout}
+        transitionLeaveTimeout={transitionLeaveTimeout}
+      >
+        {drawer}
+        <CSSTransitionGroup
+          style={contentStyle}
+          className={classnames('md-navigation-drawer-content', contentClassName, drawerType, {
+            'active': isOpen && !temporary,
+          })}
+          component="div"
+          transitionName={contentTransitionName}
+          transitionEnter={!!contentTransitionEnterTimeout}
+          transitionEnterTimeout={contentTransitionEnterTimeout}
+          transitionLeave={!!contentTransitionLeaveTimeout}
+          transitionLeaveTimeout={contentTransitionLeaveTimeout}
+        >
+          <DrawerToolbar
             isOpen={isOpen}
+            drawerType={drawerType}
+            style={toolbarStyle}
+            className={toolbarClassName}
+            temporary={temporary}
             persistent={persistent}
-            openDrawer={openDrawer}
-            menuIconClassName={menuIconClassName}
+            openDrawer={this.openDrawer}
             menuIconChildren={menuIconChildren}
+            menuIconClassName={menuIconClassName}
             title={toolbarTitle}
             children={toolbarChildren}
-            style={toolbarStyle}
-            touch={touch}
           />
           {children}
-          <Overlay isOpen={isOpen && (!persistent || persistent && touch)} onClick={closeDrawer} />
-        </div>
-      </div>
+          {overlay}
+        </CSSTransitionGroup>
+      </CSSTransitionGroup>
     );
   }
 }
